% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_data_by_date.R
\name{cut_data_by_date}
\alias{cut_data_by_date}
\title{Cut nb_sim data by calendar date}
\usage{
cut_data_by_date(data, cut_date, ...)
}
\arguments{
\item{data}{Data generated by \code{\link[=nb_sim]{nb_sim()}}.}

\item{cut_date}{Calendar time (relative to trial start) at which to censor follow-up.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A data frame with one row per subject randomized prior to \code{cut_date} containing
the truncated follow-up time (\code{tte}) and total number of observed events (\code{events}).
}
\description{
Cut nb_sim data by calendar date
}
\examples{
enroll_rate <- data.frame(rate = 20 / (5 / 12), duration = 5 / 12)
fail_rate <- data.frame(treatment = c("Control", "Experimental"), rate = c(0.5, 0.3))
dropout_rate <- data.frame(
  treatment = c("Control", "Experimental"),
  rate = c(0.1, 0.05), duration = c(100, 100)
)
sim <- nb_sim(enroll_rate, fail_rate, dropout_rate, max_followup = 2, n = 20)
cut_data_by_date(sim, cut_date = 1)
}
