\name{NBDesign-package}
\alias{NBDesign-package}
\alias{NBDesign}
\docType{package}
\title{
\packageTitle{NBDesign}
}
\description{
\packageDescription{NBDesign}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{NBDesign}
\packageIndices{NBDesign}
}
\author{
\packageAuthor{NBDesign}

Maintainer: \packageMaintainer{NBDesign}
}
%\references{Luo (2018)}
\keyword{piecewise exponential}
\keyword{various functions}
%\seealso{
%Luo (2018)}
