% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsNBCalendar.R
\name{print.gsNBsummary}
\alias{print.gsNBsummary}
\title{Print Method for gsNBsummary Objects}
\usage{
\method{print}{gsNBsummary}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gsNBsummary}.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for gsNBsummary Objects
}
\examples{
nb_ss <- sample_size_nbinom(
  lambda1 = 0.5, lambda2 = 0.3, dispersion = 0.1, power = 0.9,
  accrual_rate = 10, accrual_duration = 20, trial_duration = 24
)
gs_design <- gsNBCalendar(nb_ss, k = 3)
s <- summary(gs_design)
print(s)
}
