% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_nbinom.R
\name{print.sample_size_nbinom_summary}
\alias{print.sample_size_nbinom_summary}
\title{Print Method for sample_size_nbinom_summary Objects}
\usage{
\method{print}{sample_size_nbinom_summary}(x, ...)
}
\arguments{
\item{x}{An object of class \code{sample_size_nbinom_summary}.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for sample_size_nbinom_summary Objects
}
\examples{
x <- sample_size_nbinom(
  lambda1 = 0.5, lambda2 = 0.3, dispersion = 0.1, power = 0.8,
  accrual_rate = 10, accrual_duration = 20, trial_duration = 24
)
s <- summary(x)
print(s)

}
