% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsNBCalendar.R
\name{print.gsNBsummary}
\alias{print.gsNBsummary}
\title{Print method for gsNBsummary objects}
\usage{
\method{print}{gsNBsummary}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gsNBsummary}.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Print method for gsNBsummary objects
}
