% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_nbinom.R
\name{print.sample_size_nbinom_summary}
\alias{print.sample_size_nbinom_summary}
\title{Print method for sample_size_nbinom_summary objects}
\usage{
\method{print}{sample_size_nbinom_summary}(x, ...)
}
\arguments{
\item{x}{An object of class \code{sample_size_nbinom_summary}.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Print method for sample_size_nbinom_summary objects
}
